update
setting update slider mainpage
